requirements/base.txt
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==10.1.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.6.0  # https://github.com/evansd/whitenoise
redis==5.0.1  # https://github.com/redis/redis-py
hiredis==2.2.3  # https://github.com/redis/hiredis-py

# Django
# ------------------------------------------------------------------------------
django==4.2.7  # pyup: < 5.0  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.58.2  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.1  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==2023.10  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.4.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
django-cors-headers==4.3.1  # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular==0.26.5  # https://github.com/tfranzel/drf-spectacular
django-vue-utils==0.1.4  # https://github.com/ilikerobots/django-vue-utilities

requests~=2.31.0
pytest~=7.4.3
django-debug-toolbar~=4.2.0
